delete old fee bump context
