--- apps/coordinator/src/components/Wallet/TransactionsTab/FeeBumping/context/index.ts
@@ -1,3 +1,0 @@
-export * from "./FeeBumpContext";
-export * from "./feeBumpActions";
-export * from "./feeBumpReducer";
